syncing from original (#1)
* Update Chapter 18.

Very small changes to improve readability.

* Revamp README

* English screenshots of chap02 program

* Add link to launch article

* Fix typos chapter01.md

Small typo fixes on chapter01:
which -> whose
can not -> cannot

* Remove useless variable

* Update contributor list

* Added event listener to code example

The code example didn't have an event listener, so it just executed without the need of the button click. So I added the event listener.

* Add contributor

* grammar fix: "a URL"

I know it's minor, but URL as an abbreviation is pronounced as 'you-are-ell'. Y is a consonant, and the general rule is that 'an' is reserved for vowels, and 'a' is reserved for consonants.

* Update contributor list

* Fix #73

* Fix #104

* Fix #74

* (Try to) Fix #105

* Move environment setup chapter from appendices to intro

* Add Amazon to README

* Fix #130

* Typo fix

* Fixed a tiny, tiny typo

"request" was in source tags.

* Update copyright and add a contributor

* Update chapter24.md

Fix typo line 133

* Update contributor list